# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tiligo delivery app API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Shop {
  id: string;
  name: string;
  description?: string;
  category: string;
  address: string;
  city: string;
  phone: string;
  email: string;
  imageUrl?: string;
  rating?: number;
  reviewCount?: number;
  deliveryTime?: string;
  deliveryFee?: number;
  minOrder?: number;
  isOpen: boolean;
  createdAt: string;
}

export interface CreateShopRequest {
  name: string;
  description?: string;
  category: string;
  address: string;
  city: string;
  phone: string;
  email: string;
  password: string;
  imageUrl?: string;
  deliveryTime?: string;
  deliveryFee?: number;
  minOrder?: number;
}

export interface Product {
  id: string;
  shopId: string;
  name: string;
  description?: string;
  price: number;
  imageUrl?: string;
  category?: string;
  isAvailable: boolean;
  suggestions?: string[];
  createdAt: string;
}

export interface CreateProductRequest {
  name: string;
  description?: string;
  price: number;
  imageUrl?: string;
  category?: string;
  isAvailable?: boolean;
  suggestions?: string[];
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  accepted: "accepted",
  preparing: "preparing",
  picked_up: "picked_up",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface OrderItem {
  productId: string;
  productName: string;
  quantity: number;
  price: number;
  suggestions?: string[];
}

export interface Order {
  id: string;
  shopId: string;
  shopName: string;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  items: OrderItem[];
  totalAmount: number;
  deliveryFee?: number;
  status: OrderStatus;
  deliveryPersonId?: string;
  deliveryPersonName?: string;
  notes?: string;
  estimatedDeliveryTime?: string;
  createdAt: string;
  updatedAt: string;
}

export interface OrderItemRequest {
  productId: string;
  productName: string;
  quantity: number;
  price: number;
  suggestions?: string[];
}

export interface CreateOrderRequest {
  shopId: string;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  items: OrderItemRequest[];
  notes?: string;
}

export type UpdateOrderStatusRequestStatus =
  (typeof UpdateOrderStatusRequestStatus)[keyof typeof UpdateOrderStatusRequestStatus];

export const UpdateOrderStatusRequestStatus = {
  pending: "pending",
  accepted: "accepted",
  preparing: "preparing",
  picked_up: "picked_up",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusRequest {
  status: UpdateOrderStatusRequestStatus;
  deliveryPersonId?: string;
  deliveryPersonName?: string;
  estimatedDeliveryTime?: string;
}

export interface DeliveryPerson {
  id: string;
  name: string;
  phone: string;
  email: string;
  vehicle?: string;
  isAvailable: boolean;
  createdAt: string;
}

export interface RegisterDeliveryRequest {
  name: string;
  phone: string;
  email: string;
  password: string;
  vehicle?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type ListShopsParams = {
  category?: string;
};

export type ListOrdersParams = {
  shopId?: string;
  deliveryPersonId?: string;
  status?: string;
};