# lib/api-zod/src/generated/types/createOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tiligo delivery app API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItemRequest } from "./orderItemRequest";

export interface CreateOrderRequest {
  shopId: string;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  items: OrderItemRequest[];
  notes?: string;
}