# lib/api-zod/src/generated/types/createShopRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tiligo delivery app API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateShopRequest {
  name: string;
  description?: string;
  category: string;
  address: string;
  city: string;
  phone: string;
  email: string;
  password: string;
  imageUrl?: string;
  deliveryTime?: string;
  deliveryFee?: number;
  minOrder?: number;
}