# lib/api-zod/src/generated/types/deliveryPerson.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tiligo delivery app API
 * OpenAPI spec version: 0.1.0
 */

export interface DeliveryPerson {
  id: string;
  name: string;
  phone: string;
  email: string;
  vehicle?: string;
  isAvailable: boolean;
  createdAt: string;
}