# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tiligo delivery app API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  productId: string;
  productName: string;
  quantity: number;
  price: number;
  suggestions?: string[];
}