# lib/api-zod/src/generated/types/registerDeliveryRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tiligo delivery app API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterDeliveryRequest {
  name: string;
  phone: string;
  email: string;
  password: string;
  vehicle?: string;
}