# lib/api-zod/src/generated/types/shop.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tiligo delivery app API
 * OpenAPI spec version: 0.1.0
 */

export interface Shop {
  id: string;
  name: string;
  description?: string;
  category: string;
  address: string;
  city: string;
  phone: string;
  email: string;
  imageUrl?: string;
  rating?: number;
  reviewCount?: number;
  deliveryTime?: string;
  deliveryFee?: number;
  minOrder?: number;
  isOpen: boolean;
  createdAt: string;
}