# lib/api-zod/src/generated/types/updateOrderStatusRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tiligo delivery app API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateOrderStatusRequestStatus } from "./updateOrderStatusRequestStatus";

export interface UpdateOrderStatusRequest {
  status: UpdateOrderStatusRequestStatus;
  deliveryPersonId?: string;
  deliveryPersonName?: string;
  estimatedDeliveryTime?: string;
}